backend/pyproject.toml
[project]
name = "imagebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.2",
    "celery-types>=0.23.0",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis[hiredis]>=6.0.0",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.24",
]
